Correctly handle nullable values in unique indexes
When applying valhammer to a table which has a nullable column forming
part of a unique index, null values were being incorrectly rejected by
the validations as not unique. The behaviour of database engines in this
regard is clear - NULL values are treated as not equal.

This change ensures that the uniqueness validation only applies when all
nullable scope values have a supplied value.