Use symbolic column names for scoped unique checks
The difference in how valhammer applies validations, compared to
hand-written validations in ActiveRecord, was detected by the newer
version of shoulda-matchers which now fails because the applied scope
doesn't match the expectations.

Converting the scope columns to symbols brings this in line with how the
validations would be written without valhammer.